Update UserInterface
Closer to being ready for a demo.

Short term need to work on the Listeners  for the add_recipe_frame_maker method and work on the "new recipe" frame design.

Long term need to work on sizing. Currently all based on exact #'s for the ease of getting it all out there and working in time for a demo, but wont look right if screen is enlarged in anyway.